Split warehouse mutations
