+++ saleor/graphql/warehouse/mutations/__init__.py
@@ -1,0 +1,13 @@
+from .warehouse_create import WarehouseCreate
+from .warehouse_delete import WarehouseDelete
+from .warehouse_shipping_zone_assign import WarehouseShippingZoneAssign
+from .warehouse_shipping_zone_unassign import WarehouseShippingZoneUnassign
+from .warehouse_update import WarehouseUpdate
+
+__all__ = [
+    "WarehouseCreate",
+    "WarehouseDelete",
+    "WarehouseShippingZoneAssign",
+    "WarehouseShippingZoneUnassign",
+    "WarehouseUpdate",
+]
